requirements.txt
anyio==3.6.2
autopep8==2.0.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
fastapi==0.95.1
h11==0.14.0
idna==3.4
iniconfig==2.0.0
packaging==23.1
pluggy==1.0.0
pycodestyle==2.10.0
pydantic==1.10.7
pytest==7.3.1
python-dotenv==1.0.0
requests==2.30.0
sniffio==1.3.0
starlette==0.26.1
typing_extensions==4.5.0
urllib3==2.0.2
uvicorn==0.22.0
